refactor: Decouple the module from @appium/gulp-plugins
