+++ .mocharc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  require: ['@babel/register'],
+  forbidOnly: Boolean(process.env.CI)
+};
