+++ lib/xcode.js
@@ -105,9 +105,14 @@ async function getPathFromXcodeSelect (timeout = XCRUN_TIMEOUT) {
 const getPath = _.memoize(function getPath (timeout = XCRUN_TIMEOUT) {
   // first we try using xcode-select to find the path
   // then we try using the symlinks that Apple has by default
-  return getPathFromXcodeSelect(timeout).catch(getPathFromSymlink);
+  return (async () => {
+    try {
+      return await getPathFromXcodeSelect(timeout);
+    } catch (e) {
+      return await getPathFromSymlink(e.message);
+    }
+  })();
 });
-
 
 
 async function getVersionWithoutRetry (timeout = XCRUN_TIMEOUT) {
