+++ package.json
@@ -7,7 +7,7 @@
     "xcode"
   ],
   "version": "4.0.0",
-  "author": "appium",
+  "author": "Appium Contributors",
   "license": "Apache-2.0",
   "repository": {
     "type": "git",
@@ -16,9 +16,10 @@
   "bugs": {
     "url": "https://github.com/appium/appium-xcode/issues"
   },
-  "engines": [
-    "node"
-  ],
+  "engines": {
+    "node": ">=14",
+    "npm": ">=8"
+  },
   "main": "./build/index.js",
   "bin": {},
   "directories": {
@@ -41,32 +42,51 @@
     "teen_process": "^2.0.0"
   },
   "scripts": {
-    "clean": "rm -rf node_modules && rm -f package-lock.json && npm install",
-    "prepare": "gulp prepublish",
-    "test": "gulp once",
-    "watch": "gulp watch",
-    "mocha": "mocha",
-    "build": "gulp transpile",
-    "e2e-test": "gulp e2e-test",
-    "coverage": "gulp coveralls",
+    "build": "rm -rf build/lib/* && babel --out-dir=build/lib lib",
+    "dev": "npm run build -- --watch",
+    "lint": "eslint .",
+    "lint:fix": "npm run lint -- --fix",
+    "precommit-lint": "lint-staged",
     "precommit-msg": "echo 'Pre-commit checks...' && exit 0",
-    "lint": "gulp lint",
-    "lint:fix": "gulp eslint --fix"
+    "prepare": "npm run build",
+    "test": "mocha --exit --timeout 1m \"./test/unit/**/*-specs.js\"",
+    "e2e-test": "mocha --exit --timeout 5m \"./test/e2e/**/*-specs.js\""
   },
   "pre-commit": [
     "precommit-msg",
-    "test"
+    "precommit-lint"
   ],
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix"
+    ]
+  },
+  "prettier": {
+    "bracketSpacing": false,
+    "printWidth": 100,
+    "singleQuote": true
+  },
   "devDependencies": {
-    "@appium/gulp-plugins": "^7.0.0",
     "@appium/eslint-config-appium": "^6.0.0",
+    "@babel/cli": "^7.18.10",
+    "@babel/core": "^7.18.10",
+    "@babel/eslint-parser": "^7.18.9",
+    "@babel/plugin-transform-runtime": "^7.18.10",
+    "@babel/preset-env": "^7.18.10",
+    "@babel/register": "^7.18.9",
     "@semantic-release/git": "^10.0.1",
+    "babel-plugin-source-map-support": "^2.2.0",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.5.0",
-    "gulp": "^4.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-mocha": "^9.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "lint-staged": "^13.0.3",
     "mocha": "^10.0.0",
     "pre-commit": "^1.1.3",
+    "prettier": "^2.7.1",
     "semantic-release": "^19.0.2"
   }
 }
