--- test/.eslintrc
@@ -1,5 +1,0 @@
-{
-    "rules": {
-        "func-names": 0
-    }
-}
